Update models based on H2 scripts.  Fixed 'skills to skill' in script.  moved files to the correct folders
