+++ Curricula/src/main/java/com/revature/daos/CurriculumDao.java
@@ -1,4 +1,4 @@
-package com.revature.dao;
+package com.revature.daos;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
@@ -7,5 +7,4 @@ import com.revature.models.Curriculum;
 
 @Repository
 public interface CurriculumDao extends JpaRepository<Curriculum,Integer> {
-
 }
