+++ Curricula/src/main/java/com/revature/models/Curriculum.java
@@ -2,14 +2,8 @@ package com.revature.models;
 
 import java.util.List;
 
-import javax.persistence.Column;
-import javax.persistence.Entity;
-import javax.persistence.FetchType;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
-import javax.persistence.JoinColumn;
-import javax.persistence.OneToMany;
+import javax.persistence.*;
+
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 
 @Entity
@@ -24,8 +18,11 @@ public class Curriculum {
 	@Column(name="curriculum_name")
 	private String curriculumName;
 	
-	@OneToMany(fetch = FetchType.EAGER)
-	@JoinColumn(name="skill_id")
+	@ManyToMany
+	@JoinTable(
+		name = "curriculum_skill",
+		joinColumns = @JoinColumn(name = "curriculum_id"),
+		inverseJoinColumns = @JoinColumn(name = "skill_id"))
 	private List<Skill> skills;
 
 	public Curriculum() {
