+++ Curricula/src/main/java/com/revature/models/Visualization.java
@@ -3,14 +3,7 @@ package com.revature.models;
 import java.awt.Color;
 import java.util.List;
 
-import javax.persistence.Column;
-import javax.persistence.Entity;
-import javax.persistence.FetchType;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
-import javax.persistence.JoinColumn;
-import javax.persistence.OneToMany;
+import javax.persistence.*;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 
@@ -26,8 +19,11 @@ public class Visualization {
 	@Column(name = "visualization_name")
 	private String visualizationName;
 
-	@OneToMany(fetch = FetchType.EAGER)
-	@JoinColumn(name="curriculum_id")
+	@ManyToMany
+	@JoinTable(
+			name = "visualization_curriculum",
+			joinColumns = @JoinColumn(name = "visualization_id"),
+			inverseJoinColumns = @JoinColumn(name = "curriculum_id"))
 	private List<Curriculum> curriculum;
 
 	public Visualization() {
