+++ Curricula/src/main/java/com/revature/services/SkillServicesImpl.java
@@ -8,7 +8,9 @@ import org.springframework.beans.factory.annotation.Autowired;
 
 import com.revature.daos.SkillDao;
 import com.revature.models.Skill;
+import org.springframework.stereotype.Service;
 
+@Service
 public class SkillServicesImpl implements SkillServices {
 	
 	private SkillDao sd;
